requirements.txt
aiogram
numpy
pandas
python-dotenv
ta
torch
t-tech-investments --index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
